+++ pkg/querier/worker/frontend_processor_test.go
@@ -12,6 +12,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
+	"github.com/cortexproject/cortex/pkg/util/limiter"
 	"github.com/cortexproject/cortex/pkg/util/test"
 )
 
@@ -25,7 +26,7 @@ func TestRecvFailDoesntCancelProcess(t *testing.T) {
 	require.NoError(t, err)
 
 	cfg := Config{}
-	mgr := newFrontendProcessor(cfg, nil, log.NewNopLogger())
+	mgr := newFrontendProcessor(cfg, &limiter.NoOpMemoryLimiter{}, nil, log.NewNopLogger())
 	running := atomic.NewBool(false)
 	go func() {
 		running.Store(true)